LayoutTests/platform/gtk/imported/w3c/web-platform-tests/wpt_internal/html/canvas/drawElementImage/hit-test/selection-expected.txt

hello world

FAIL Double-clicking on canvas should be able to modify selection. assert_equals: #target text should be selected by double-click. expected Text node "hello world" but got null
